simplify the way  _AewbHandle is factored out of linux_aewb_module.c
